+++ App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { NavigationContainer } from "@react-navigation/native";
 import { createNativeStackNavigator } from "@react-navigation/native-stack";
-import Home from "./screens/Home";
-import Individual from "./screens/Individual";
+import Home from "./screens/Welcome";
+import Individual from "./screens/Home";
 import SingleGroup from "./screens/SingleGroup";
 
 export default function App() {
